Fixed conversion warnings

+++ Voyager/Source/Voyager/Math/Vector/Rotation.h
@@ -8,7 +8,7 @@ namespace VE
 		struct YawPitchRoll
 		{
 			YawPitchRoll() : yaw(0.0), pitch(0.0), roll(0.0) {}
-			YawPitchRoll(double y, double p, double r) : yaw(y), pitch(p), roll(r) {}
+			YawPitchRoll(float y, float p, float r) : yaw(y), pitch(p), roll(r) {}
 			float yaw;
 			float pitch;
 			float roll;
